Reorganize project-types code
