+++ src/util.rs
@@ -1,4 +1,3 @@
-use crate::cmd::{CommandOutputError, LoggedCommand};
 use chrono::Datelike;
 use fs_err::{create_dir_all, read_dir, remove_dir};
 use rangemap::RangeInclusiveSet;
@@ -294,40 +293,6 @@ impl DirWithEntries {
     fn pop_front(&mut self) -> Option<(PathBuf, FileType)> {
         self.entries.pop_front()
     }
-}
-
-pub(crate) fn locate_project(workspace: bool) -> Result<PathBuf, LocateError> {
-    let mut cmd = LoggedCommand::new("cargo");
-    cmd.arg("locate-project");
-    if workspace {
-        cmd.arg("--workspace");
-    }
-    let output = cmd.check_output()?;
-    let location = serde_json::from_str::<LocateProject<'_>>(&output)?;
-    if !location.root.is_absolute() {
-        return Err(LocateError::InvalidPath(location.root.into()));
-    }
-    if location.root.parent().is_some() {
-        Ok(location.root.into())
-    } else {
-        Err(LocateError::InvalidPath(location.root.into()))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq)]
-struct LocateProject<'a> {
-    #[serde(borrow)]
-    root: std::borrow::Cow<'a, Path>,
-}
-
-#[derive(Debug, Error)]
-pub(crate) enum LocateError {
-    #[error("could not get project root from cargo")]
-    Command(#[from] CommandOutputError),
-    #[error("could not deserialize `cargo locate-project` output")]
-    Deserialize(#[from] serde_json::Error),
-    #[error("manifest path is absolute or parentless: {}", .0.display())]
-    InvalidPath(PathBuf),
 }
 
 pub(crate) fn workspace_tag_prefix(package_name: &str) -> String {
